pyproject.toml
[project]
name = "resnet-ibn-reid-attr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lightning>=2.5.3",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "rich>=14.1.0",
    "tensorboard>=2.20.0",
    "tensorboardx>=2.6.4",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "torchmetrics>=1.0.0",
]
